requirements_base.txt
otree-core==0.5.13
Django==1.8.8 # for heroku, needs to be explicitly in requirements file
